requirements.txt
# annotated-doc==0.0.4
# annotated-types==0.7.0
# anyio==4.12.1
# certifi==2026.1.4
# charset-normalizer==3.4.4
# click==8.3.1
# colorama==0.4.6
# contourpy==1.3.3
# cycler==0.12.1
# dill==0.4.1
# ensemble-boxes==1.0.9
# fastapi==0.128.7
# filelock==3.20.3
# fonttools==4.61.1
# fsspec==2026.2.0
# h11==0.16.0
# idna==3.11
# Jinja2==3.1.6
# kiwisolver==1.4.9
# llvmlite==0.46.0
# MarkupSafe==3.0.3
# matplotlib==3.10.8
# mpmath==1.3.0
# networkx==3.6.1
# numba==0.63.1
# numpy==2.3.5
# opencv-python==4.13.0.92
# packaging==26.0
# pandas==3.0.0
# pillow==12.1.1
# polars==1.38.1
# polars-runtime-32==1.38.1
# psutil==7.2.2
# pydantic==2.12.5
# pydantic_core==2.41.5
# pyparsing==3.3.2
# python-dateutil==2.9.0.post0
# python-multipart==0.0.22
# PyYAML==6.0.3
# requests==2.32.5
# scipy==1.17.0
# setuptools==82.0.0
# six==1.17.0
# starlette==0.52.1
# sympy==1.14.0
# torch==2.10.0
# torchvision==0.25.0
# typing-inspection==0.4.2
# typing_extensions==4.15.0
# tzdata==2025.3
# ultralytics==8.4.14
# ultralytics-thop==2.0.18
# urllib3==2.6.3
# uvicorn==0.40.0
--extra-index-url https://download.pytorch.org/whl/cpu

annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dill==0.4.1
ensemble-boxes==1.0.9
fastapi==0.128.7
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.2.0
h11==0.16.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
llvmlite==0.46.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numba==0.63.1
numpy==2.3.5
opencv-python==4.13.0.92
packaging==26.0
pandas==3.0.0
pillow==12.1.1
polars==1.38.1
polars-runtime-32==1.38.1
psutil==7.2.2
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
scipy==1.17.0
setuptools==82.0.0
six==1.17.0
starlette==0.52.1
sympy==1.14.0
torch==2.10.0
torchvision==0.25.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.14
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.40.0
streamlit
altair==4.2.2
